Update for RDFLib 4.1.1
Adds RDF 1.1 feature support, RDFa, Microdata, and TriG parsers, HTML5
handling.

The tools originally provided by python-rdfextras are now provided as part of
python-rdflib itself, so we add an Obsoletes line accordingly.

We also make SPARQLWrapper an "extras_require" for 4.1; per upstream discussion,
SPARQLWrapper can be an optional dependency for 4.1 but will be made mandatory
for 4.2. This gives us time to ensure SPARQLWrapper is packaged while providing
all of the rest of the bug fixes and additional functionality in RDFLib 4.1.

Signed-off-by: Dan Scott <dan@coffeecode.net>