Update for RDFLib 4.1.0
Update the SPEC to reflect RDFLib 4.1.0. Includes a patch (as
submitted to upstream) to skip one test, similar to how other
tests are skipped for the same condition.

NOTE: The rdfpipe executable is now shipped as part of RDFLib
core, so the current spec as it stands will conflict with
python-rdfextras (which also ships rdfpipe) until
python-rdfextras can be converted into an empty package that
simply requires python-rdflib and then dropped entirely.

Signed-off-by: Dan Scott <dan@coffeecode.net>